+++ src/main/java/at/hannibal2/skyhanni/config/features/mining/nucleus/CrystalHighlighterColorConfig.java
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.config.features.mining;
+package at.hannibal2.skyhanni.config.features.mining.nucleus;
 
 import com.google.gson.annotations.Expose;
 import io.github.notenoughupdates.moulconfig.annotations.ConfigEditorColour;
