+++ src/main/java/at/hannibal2/skyhanni/config/features/mining/nucleus/CrystalHighlighterConfig.java
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.config.features.mining;
+package at.hannibal2.skyhanni.config.features.mining.nucleus;
 
 import at.hannibal2.skyhanni.config.FeatureToggle;
 import com.google.gson.annotations.Expose;
